toplevel language model class
